+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.4] - 2022-06-28
+
+### Added
+
+- A new attached dataset of GWAS summary statistics for standing height from the fastGWA database.
+
+### Changed
+
+- Updated the data harmonization method in `GWADataLoader` to ensure that all data sources 
+have the same set of chromosomes.
+- Bug fixes in `SumStatsTable`, `GWASimulator`, and `GWADataLoader`.
+
 ## [0.0.3] - 2022-06-26
 
 ### Added